refactor: 롯데마트 소켓 전송 계층 분리
